Force SDR and SIR to be in the idle state after running and passing through the capture state.

+++ desktop_client/src/com/doublecheck/bstworkbench/compiler/commands/SeltapCommand.java
@@ -63,7 +63,7 @@ public class SeltapCommand extends Command {
     @Override
     public List<Instruction> getInstructions() {
         List<Instruction> ret = new ArrayList<Instruction>(1);
-        ret.add(new Instruction(Command.SELTAP, 1, (long)tapNumber));
+        ret.add(new Instruction(Command.SELTAP, (byte)1, (long)tapNumber));
         return ret;
     }
